Add a sample-and-match stage.
The sample-and-match stage randomly samples the set of all families
and the set of daughters. In the case of the 2GG families it then
samples the daughters. If the daughter from the family side and
from the daughter side match, then the family is output. Otherwise
the loop repeats.

This is done n times, and the statistics calculated.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>